Typo and code fix
The IIFE example was missing an inner function expression that wraps the console log statement and gets returned so that it can get pushed to the `fns` array in order for it to be accessed and executed in the subsequent for-loop. Before, the code would produce the error "TypeError: fn is not a function".